requirements.txt
python-telegram-bot
python-dotenv
pyserial
schedule
